refactor(web): update action dir
